requirements.txt
cryptography
requests
google-auth
